Have ClassContext load a class from the ClassReader
This change is done in order to prevent the JVM from running unnecessary verifications and all that JVM stuff that might interfere with modifying the class file. Furthermore, loading from the ClassReader allows us to preload all the method body so that we have a more standard way of loading and saving classes later on. This adds a few classes, including the MethodBody, which serves to store all the metadata of a method that isn't otherwise in the header. Currently, this method body doesn't store annotations, which is fine, because we will implement that later on.

Furthermore, the IndexHashSet is used in order to preserve member order of fields and methods within our class context body.