Implement stack map frame.
The stack map frame will simulate the changes done to a stack so to
enforce a strict type-safe operation stack and locals. Currently,
it is likely that the user will still have to manually specify stack
frames at any instruction i where i is the target of any jump
instruction, or is the beginning of a catch clause. This is done in
order to prevent deep complex loops that can be constructed, to cut
down on time on the verification system.

As of this commit, the tests for stack map frames have not been
implemented yet, nor has this been used in the main code parsing
stuff.